requirements.txt
Django==5.2.3
djangorestframework==3.15.2
mysqlclient==2.2.7
openpyxl==3.1.5
pandas==2.2.3
requests==2.32.3
reportlab==4.4.1
PyPDF2==3.0.1